#38 Refactor codes
