+++ pom.xml
@@ -6,7 +6,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>plugin.ecpg</groupId>
   <artifactId>plugin.ecpg</artifactId>
-  <version>0.2.4</version>
+  <version>0.2.5</version>
   <packaging>eclipse-plugin</packaging>
   <properties>
     <tycho.version>4.0.12</tycho.version>
