+++ src/plugin/ecpg/parser/scanner/ECPGScanner.java
@@ -13,6 +13,7 @@ import org.eclipse.cdt.core.parser.IToken;
 import org.eclipse.cdt.core.parser.IncludeFileContentProvider;
 import org.eclipse.cdt.core.parser.ParserLanguage;
 
+import plugin.common.parser.scanner.ExecSqlPosition;
 import plugin.common.parser.scanner.OverrideToken;
 import plugin.common.parser.scanner.PluginScanner;
 import plugin.common.parser.scanner.PluginTokenUtil;
@@ -88,9 +89,7 @@ public class ECPGScanner extends PluginScanner {
 					immediateAfterVariable = true;
 					// Modify tokens to be treated as variables.
 					tokens().add(nextToken);
-					tokens().add(PluginTokenUtil.createAssignToken(this, nextToken));
-					tokens().add(PluginTokenUtil.createLiteralToken(this, nextToken, ""));
-					tokens().add(PluginTokenUtil.createSemiToken(this, nextToken));
+					PluginTokenUtil.addAssignLiteralTokens(this, nextToken);
 					continue;
 				}
 				if (":".equals(nextToken.toString())) {
@@ -99,9 +98,7 @@ public class ECPGScanner extends PluginScanner {
 					immediateAfterVariable = true;
 					// Modify tokens to be treated as variables.
 					tokens().add(nextToken);
-					tokens().add(PluginTokenUtil.createAssignToken(this, nextToken));
-					tokens().add(PluginTokenUtil.createIntegerToken(this, nextToken, "0"));
-					tokens().add(PluginTokenUtil.createSemiToken(this, nextToken));
+					PluginTokenUtil.addAssignIntegerTokens(this, nextToken);
 					colonStarted = true;
 					immediateAfterComma = false;
 					continue;
@@ -121,9 +118,7 @@ public class ECPGScanner extends PluginScanner {
 					immediateAfterVariable = true;
 					// Modify tokens to be treated as variables.
 					tokens().add(nextToken);
-					tokens().add(PluginTokenUtil.createAssignToken(this, nextToken));
-					tokens().add(PluginTokenUtil.createIntegerToken(this, nextToken, "0"));
-					tokens().add(PluginTokenUtil.createSemiToken(this, nextToken));
+					PluginTokenUtil.addAssignIntegerTokens(this, nextToken);
 					continue;
 				}
 				if (";".equals(nextToken.toString())) {
@@ -148,41 +143,11 @@ public class ECPGScanner extends PluginScanner {
 		dummyToken.setEndOffset(nextToken.getOffset() - 1);
 		PluginTokenUtil.addSqlcaTokensEcpg(this, dummyToken);
 		// Initialize sqlca struct.
-		tokens().add(PluginTokenUtil.createStructToken(this, dummyToken));
-		tokens().add(PluginTokenUtil.createIdentifierToken(this, dummyToken, "sqlca"));
-		tokens().add(PluginTokenUtil.createIdentifierToken(this, dummyToken, "sqlca"));
-		tokens().add(PluginTokenUtil.createSemiToken(this, dummyToken));
+		PluginTokenUtil.addSqlcaStructTokens(this, dummyToken);
 		// SQLSTATE
-		tokens().add(PluginTokenUtil.createCharToken(this, dummyToken));
-		tokens().add(PluginTokenUtil.createIdentifierToken(this, dummyToken, "SQLSTATE"));
-		tokens().add(PluginTokenUtil.createSemiToken(this, dummyToken));
+		PluginTokenUtil.addSqlStateTokens(this, dummyToken);
 		// SQLCODE
-		tokens().add(PluginTokenUtil.createLongToken(this, dummyToken));
-		tokens().add(PluginTokenUtil.createIdentifierToken(this, dummyToken, "SQLCODE"));
-		tokens().add(PluginTokenUtil.createSemiToken(this, dummyToken));
+		PluginTokenUtil.addSqlCodeTokens(this, dummyToken);
 	}
 
-	public class ExecSqlPosition {
-		int from;
-		int to;
-
-		public ExecSqlPosition(int from, int to) {
-			this.from = from;
-			this.to = to;
-		}
-
-		public int getFrom() {
-			return from;
-		}
-
-		public int getTo() {
-			return to;
-		}
-
-		@Override
-		public String toString() {
-			return from + ":" + to;
-		}
-
-	}
 }
